refactor: move module declarations for non-route components to theme-classic
